Change localize ID and fix registration
