+++ src/sql/workbench/parts/queryHistory/electron-browser/queryHistory.contribution.ts
@@ -7,15 +7,5 @@ import { Registry } from 'vs/platform/registry/common/platform';
 import { IWorkbenchContributionsRegistry, Extensions as WorkbenchExtensions } from 'vs/workbench/common/contributions';
 import { LifecyclePhase } from 'vs/platform/lifecycle/common/lifecycle';
 import { QueryHistoryWorkbenchContribution } from 'sql/workbench/parts/queryHistory/electron-browser/queryHistory';
-import { CommandsRegistry, ICommandService } from 'vs/platform/commands/common/commands';
-import { IQueryHistoryService } from 'sql/platform/queryHistory/common/queryHistoryService';
 
 Registry.as<IWorkbenchContributionsRegistry>(WorkbenchExtensions.Workbench).registerWorkbenchContribution(QueryHistoryWorkbenchContribution, LifecyclePhase.Restored);
-
-CommandsRegistry.registerCommand({
-	id: 'queryHistory.clear',
-	handler: (accessor) => {
-		const queryHistoryService = accessor.get(IQueryHistoryService);
-		queryHistoryService.clearQueryHistory();
-	}
-});
